Create teacher class and all tests pass
