+++ senior_teacher.rb
@@ -1,10 +1,12 @@
 require_relative 'person'
 require_relative 'Highfive'
+require_relative 'teacher'
 
-class SeniorTeacher < Person
+class SeniorTeacher < Teacher
   attr_reader :salary, :performance_rating, :target_raise
   
    include Highfive
+   RATING = 90
 
   def initialize(options={})
     super
@@ -13,10 +15,7 @@ class SeniorTeacher < Person
     @target_raise = 1000
   end
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  
 
   def teach_stuff
     response = ""
@@ -26,28 +25,7 @@ class SeniorTeacher < Person
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
-
-  def lead_training_session
+   def lead_training_session
     puts "Hey newbie!  Here are some common pitfalls.  Don't fall in them!"
   end
 end
