+++ spec/senior_teacher_spec.rb
@@ -1,6 +1,7 @@
 require_relative '../senior_teacher'
 require_relative '../person'
 require_relative '../Highfive'
+require_relative '../teacher'
 
 describe SeniorTeacher do
   let(:jan) { SeniorTeacher.new({name: "Jan", age: 35}) }
